Extract methods from AddParticipant#add_participant action
This change has no behavioural effect, but is a pure refactor, extracting two
methods from the `#add_participant` action.

The motivation is to allow custom behaviour when a participant is added to an
experiment (for example, logging the enrolment with an external logging
framework).

Previously, this would entail reimplementing the whole `#add_participant`
action to add any custom behaviour. Now, just the `#add_participant_choice`
method needs to be overridden:

    class VanityController < Application Controller
      include Vanity::Rails::AddParticipant

      private

      def add_participant_choice(experiment, alternative)
        super
        # custom behaviour here
      end
    end