Resetting experiment when collecting=false removes participants
Currently, when Vanity.playground.collecting is false, participant information
is stored in the @showing ivar on the experiment instance, which is stored
globally on Vanity.playground.

In a test environment, this means that assignments for experiments can persist
between test examples. When collecting=true, the `AbTest#reset` method would
be used to reset the participants of an experiment, but this method is a no-op
when collecting=false.

This ensures that the `#reset` method cleans up experiments when
collecting=false.