Fix Rails 6 deprecation warnings
During a recent Rails 6 upgrade I noticed that the `ahoy_js` view helper
was emitting a deprecation warning. It turns out this was due to the gem
calling `render file:` with incomplete file paths (they were missing the
file extension) [1].

In Rails 6 [2], this option now requires that you provide the full path
to the file.

In order to fix this these deprecation warnings I have updated all calls
to `Vanity.template` so that they now include the file extension.

Ideally I would have liked to abstract the code handling extensions down
into `Vanity.template` but since this extension can vary I thought it
was best to be explicit and just pass it in as part of the filename.

[1] https://github.com/rails/rails/blob/03d32ca23b80e74bb3bc385a29c8f44bb60a7b8c/actionview/lib/action_view/renderer/template_renderer.rb#L29
[2] https://github.com/rails/rails/commit/c7820d8124c854760a4d288334f185de2fb99446