Add note about retiring Vanity to README
Trello: https://trello.com/c/CeY5wK3J

We are no longer using Vanity for A/B testing in the main futurelearn
app [1], having recently introduced Optimizely as a replacement.

In view of that, let's archive the `futurelearn/vanity` repo (our fork
of the vanity gem) and add a note to the README to explain why.

[1] https://github.com/futurelearn/futurelearn/pull/8695